Renaming constants file to settings.
