+++ src/automowerPlatform.ts
@@ -5,7 +5,7 @@ import {
 
 import { AutomowerAccessory, AutomowerContext } from './automowerAccessory';
 import { PlatformContainer, PlatformContainerImpl } from './primitives/platformContainer';
-import { PLATFORM_NAME, PLUGIN_ID } from './constants';
+import { PLATFORM_NAME, PLUGIN_ID } from './settings';
 import { SettingsEvent, StatusEvent } from './events';
 import { AccessTokenManager, AccessTokenManagerImpl } from './services/authentication/accessTokenManager';
 import { EventStreamService, EventStreamServiceImpl } from './services/automower/eventStreamService';
