+++ src/index.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 
 import { API } from 'homebridge';
 import { AutomowerPlatform } from './automowerPlatform';
-import { PLATFORM_NAME } from './constants';
+import { PLATFORM_NAME } from './settings';
 
 export default (api: API) => { 
     api.registerPlatform(PLATFORM_NAME, AutomowerPlatform);
